requirements.txt
requests==2.32.4
pyyaml==6.0.2
dotenv==0.9.9
